Remove pseudo “search” lib
It’s quite specific to one use case (hardcoding “AllFolders”),
so shouldn’t pretend to be a library function.
